requirements.txt
fastapi
uvicorn
python-dotenv
requests
edge-tts
openai
whisper
ffmpeg-python
google-cloud-speech
